Refactor entire app interaction.
Every component behavior was handled by a higher entity. Every update
and draw was inlined by a state verificator. This refactor aims to limit
that. We're focusing on giving has much responsabilities to lower
components. The event folder will disappear and Sprite struct will be
part of the component folder.
Each component aims to have its own Update and Draw function so that
menu, lobby & game can only tell its components to update or draw
without micromanaging every movement that should happen.

By refactoring this way, we hope to facilitate server game state
handling. It'll be easier to update the game after each message received
without having to question every move done by user but by simply running
it.

A second objective of this refactor is to adding way more parameters
check and unvalid values passed between components.

Asset loading will, in the first place, stay as it is. It might be
easier to move all this asset loading directly in components. This logic
is quite in place at some level; we aim to reduce the view folder into a
single util files with generic function called by struct init.